+++ packages/hono-takibi/src/generators/hono/openapi/generate-zod-openapi-hono.ts
@@ -1,8 +1,8 @@
-import type { Config } from '../../config'
-import type { OpenAPISpec } from '../../types'
-import { generateComponentsCode } from '../openapi/components/generate-components-code'
-import { generateRouteCode } from '../openapi/paths/generate-route-code'
-import { generateTypesCode } from '../types/generate-types-code'
+import type { Config } from '../../../config'
+import type { OpenAPISpec } from '../../../types'
+import { generateComponentsCode } from './component/generate-components-code'
+import { generateRouteCode } from '../../openapi/paths/generate-route-code'
+import { generateTypesCode } from '../../types/generate-types-code'
 
 /**
  * Import statement for Hono's zod-openapi package
