+++ packages/hono-takibi/src/vite-plugin/vite-mode.ts
@@ -1,7 +1,7 @@
 import SwaggerParser from '@apidevtools/swagger-parser'
 import fs from 'node:fs'
 import path from 'node:path'
-import { generateZodOpenAPIHono } from '../generators/hono/generate-zod-openapi-hono'
+import { generateZodOpenAPIHono } from '../generators/hono/openapi/generate-zod-openapi-hono'
 import type { OpenAPISpec } from '../types'
 import type { Config } from '../config'
 import { getConfig } from '../config'
